<commit message>Refactor proto paths and update build settings
