+++ .gitignore
@@ -34,4 +34,7 @@ report.[0-9]_.[0-9]_.[0-9]_.[0-9]_.json
 .DS_Store
 build
 /proto/**
+src/proto/*.ts
+src/proto/**/*.ts
+
 src/assets/proto/**
