+++ build.ts
@@ -14,6 +14,7 @@ await Bun.build({
   entrypoints: ['./src/index.ts'],
   outdir: './dist',
   target:"bun",
+  sourcemap:"external",
   plugins: [
     // dts()
   ],
