+++ proto-gen.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
 bun run scripts/patch-protobufjs.js
-bunx proto-loader-gen-types --grpcLib=@grpc/grpc-js --outDir=proto/ proto/*.proto
+bunx proto-loader-gen-types --grpcLib=@grpc/grpc-js --outDir=src/proto/ src/proto/*.proto
 
 echo "✓ Done"
