--- proto/node.proto
@@ -1,53 +1,0 @@
-syntax = "proto3";
-package cex_broker; 
-
-message ActionRequest {
-  Action action = 1;                        // The CCXT method to call (e.g., "fetchBalance", "createOrder")
-  map<string, string> payload = 2;          // Parameters to pass to the CCXT method
-  string cex = 3;     // CEX identifier (e.g., "binance", "bybit")
-  string symbol = 4;  // Optional: trading pair symbol if needed
-}
-
-message ActionResponse {
-  string result = 2;       // JSON string of the result data
-}
-
-message SubscribeRequest {
-  string cex = 1;                          // CEX identifier (e.g., "binance", "bybit")
-  string symbol = 2;                        // Trading pair symbol (e.g., "BTC/USDT")
-  SubscriptionType type = 3;                // Type of subscription (orderbook, trades, etc.)
-  map<string, string> options = 4;          // Additional subscription options
-}
-
-message SubscribeResponse {
-  string data = 1;                         // JSON string of the streaming data
-  int64 timestamp = 2;                     // Unix timestamp of the data
-  string symbol = 3;                       // Trading pair symbol
-  SubscriptionType type = 4;               // Type of subscription
-}
-
-enum SubscriptionType {
-  ORDERBOOK = 0;                           // Order book updates
-  TRADES = 1;                              // Recent trades
-  TICKER = 2;                              // Ticker information
-  OHLCV = 3;                               // OHLCV candlestick data
-  BALANCE = 4;                             // Balance updates
-  ORDERS = 5;                              // Order updates
-}
-
-service cex_service {
-  rpc ExecuteAction(ActionRequest) returns (ActionResponse);
-  rpc Subscribe(SubscribeRequest) returns (stream SubscribeResponse);
-}
-
-// Mode enum for price direction
-enum Action {
-  NoAction=0;
-  Deposit = 1;
-  Transfer = 2;
-  CreateOrder= 3;
-  GetOrderDetails=4;
-  CancelOrder=5;
-  FetchBalance=6;
-  FetchDepositAddresses=7;
-}
