+++ src/client.dev.ts
@@ -1,17 +1,26 @@
 import * as grpc from "@grpc/grpc-js";
 import * as protoLoader from "@grpc/proto-loader";
-import type { ProtoGrpcType } from "../proto/node";
-import { Action } from "../proto/cex_broker/Action";
+import path from "path";
+import { fileURLToPath } from "url";
+import type { ProtoGrpcType } from "./proto/node";
+import { Action } from "./proto/cex_broker/Action";
 import { config } from "dotenv";
 import { log } from "./helpers/logger";
 // import CEXBroker from ".";
 import CEXBroker from "../dist/index";
 import { loadPolicy } from "./helpers";
 
-const PROTO_FILE = "./proto/node.proto";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+
+// Safe absolute path to proto
+const protoPath = path.join(__dirname, ".", "proto", "node.proto");
+
 const port = 8086;
 
-const packageDef = protoLoader.loadSync(PROTO_FILE);
+const packageDef = protoLoader.loadSync(protoPath);
 const grpcObj = grpc.loadPackageDefinition(
 	packageDef,
 ) as unknown as ProtoGrpcType;
@@ -44,15 +53,6 @@ client.waitForReady(deadline, (err) => {
 });
 
 function onClientReady() {
-	// // Test ExecuteAction for balance
-	// client.executeAction({ cex: "bybit", symbol: "USDT",payload:{},action: Action.FetchBalance },metadata, (err, result) => {
-	// 	if (err) {
-	// 		log.error({ err });
-	// 		return;
-	// 	}
-	// 	log.info("ExecuteAction Balance Result:", { result });
-	// });
-
 	// Test ExecuteAction for balance
 	client.executeAction(
 		{
@@ -71,41 +71,4 @@ function onClientReady() {
 		},
 	);
 
-	// // Test Subscribe for balance streaming
-	// log.info("Starting balance subscription test...");
-	// const subscribeCall = client.subscribe(
-	// 	{
-	// 		cex: "bybit",
-	// 		symbol: "BTC/USDT",
-	// 		type: SubscriptionType.TICKER,
-	// 		options: {},
-	// 	},
-	// 	metadata,
-	// );
-
-	// // Handle incoming stream data
-	// subscribeCall.on("data", (response) => {
-	// 	log.info("Balance Subscription Update:", {
-	// 		symbol: response.symbol,
-	// 		type: response.type,
-	// 		data: JSON.parse(response.data),
-	// 	});
-	// });
-
-	// // Handle stream end
-	// subscribeCall.on("end", () => {
-	// 	log.info("Balance subscription stream ended");
-	// });
-
-	// // Handle stream errors
-	// subscribeCall.on("error", (error) => {
-	// 	log.error("Balance subscription stream error:", error);
-	// });
-
-	// // Keep the subscription alive for 30 seconds
-	// setTimeout(() => {
-	// 	log.info("Closing balance subscription after 30 seconds");
-	// 	// For server-side streaming, we don't need to call end() on the client
-	// 	// The server will handle the stream lifecycle
-	// }, 3000000);
 }
